Make client of type `size_t` (#322)

+++ pyvrp/cpp/Solution.h
@@ -36,7 +36,7 @@ namespace pyvrp
  */
 class Solution
 {
-    using Client = int;
+    using Client = size_t;
     using VehicleType = size_t;
 
 public:
